+++ hpvsim/sim.py
@@ -734,7 +734,7 @@ class Sim(hpb.BaseSim):
         # Determine if there are any reactivated infections on this timestep
         for g in range(ng):
             latent_inds = hpu.true(people.latent[g,:])
-            if len(latent_inds): # TODO: add in increased risk of reactivation for immunocompromised
+            if len(latent_inds):
                 age_inds = np.digitize(people.age[latent_inds], self['hpv_reactivation']['age_cutoffs'])-1 # convert ages to indices
                 reactivation_probs = self['hpv_reactivation']['hpv_reactivation_probs'][age_inds]